+++ library/dtmd-library.h
@@ -24,6 +24,8 @@
 #include <dtmd.h>
 #include <dtmd-misc.h>
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -62,7 +64,7 @@ typedef struct dtmd_device
 {
 	char *path;
 	dtmd_removable_media_type_t type;
-	unsigned int partitions_count;
+	size_t partitions_count;
 	dtmd_partition_t **partition;
 } dtmd_device_t;
 
@@ -81,26 +83,26 @@ dtmd_t* dtmd_init(dtmd_callback_t callback, void *arg, dtmd_result_t *result);
 void dtmd_deinit(dtmd_t *handle);
 
 // timeout is in milliseconds, negative for infinite
-dtmd_result_t dtmd_enum_devices(dtmd_t *handle, int timeout, unsigned int *device_count, dtmd_device_t ***result, unsigned int *stateful_device_count, dtmd_stateful_device_t ***result_stateful);
+dtmd_result_t dtmd_enum_devices(dtmd_t *handle, int timeout, size_t *device_count, dtmd_device_t ***result, size_t *stateful_device_count, dtmd_stateful_device_t ***result_stateful);
 dtmd_result_t dtmd_list_device(dtmd_t *handle, int timeout, const char *device_path, dtmd_device_t **result);
 dtmd_result_t dtmd_list_partition(dtmd_t *handle, int timeout, const char *partition_path, dtmd_partition_t **result);
 dtmd_result_t dtmd_list_stateful_device(dtmd_t *handle, int timeout, const char *device_path, dtmd_stateful_device_t **result);
 dtmd_result_t dtmd_mount(dtmd_t *handle, int timeout, const char *path, const char *mount_options);
 dtmd_result_t dtmd_unmount(dtmd_t *handle, int timeout, const char *path);
-dtmd_result_t dtmd_list_supported_filesystems(dtmd_t *handle, int timeout, unsigned int *supported_filesystems_count, const char ***supported_filesystems_list);
-dtmd_result_t dtmd_list_supported_filesystem_options(dtmd_t *handle, int timeout, const char *filesystem, unsigned int *supported_filesystem_options_count, const char ***supported_filesystem_options_list);
+dtmd_result_t dtmd_list_supported_filesystems(dtmd_t *handle, int timeout, size_t *supported_filesystems_count, const char ***supported_filesystems_list);
+dtmd_result_t dtmd_list_supported_filesystem_options(dtmd_t *handle, int timeout, const char *filesystem, size_t *supported_filesystem_options_count, const char ***supported_filesystem_options_list);
 int dtmd_is_state_invalid(dtmd_t *handle);
 
 // if error is command_failed, detailed error code can be with following function
 dtmd_error_code_t dtmd_get_code_of_command_fail(dtmd_t *handle);
 
-void dtmd_free_devices_array(dtmd_t *handle, unsigned int device_count, dtmd_device_t **devices);
+void dtmd_free_devices_array(dtmd_t *handle, size_t device_count, dtmd_device_t **devices);
 void dtmd_free_device(dtmd_t *handle, dtmd_device_t *device);
 void dtmd_free_partition(dtmd_t *handle, dtmd_partition_t *partition);
-void dtmd_free_stateful_devices_array(dtmd_t *handle, unsigned int stateful_device_count, dtmd_stateful_device_t **stateful_devices);
+void dtmd_free_stateful_devices_array(dtmd_t *handle, size_t stateful_device_count, dtmd_stateful_device_t **stateful_devices);
 void dtmd_free_stateful_device(dtmd_t *handle, dtmd_stateful_device_t *stateful_device);
-void dtmd_free_supported_filesystems_list(dtmd_t *handle, unsigned int supported_filesystems_count, const char **supported_filesystems_list);
-void dtmd_free_supported_filesystem_options_list(dtmd_t *handle, unsigned int supported_filesystem_options_count, const char **supported_filesystem_options_list);
+void dtmd_free_supported_filesystems_list(dtmd_t *handle, size_t supported_filesystems_count, const char **supported_filesystems_list);
+void dtmd_free_supported_filesystem_options_list(dtmd_t *handle, size_t supported_filesystem_options_count, const char **supported_filesystem_options_list);
 
 #ifdef __cplusplus
 }
